feat: rename session filename to differentiate from the other context module
